backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==3.2.16
django-celery-beat==2.5.0
django-mail-templated==2.6.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
kombu==5.3.5
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
python-dotenv==1.0.0
psycopg2-binary==2.9.3
django-redis